envelope patch for short notes
while playing with one of the synths, I noticed that when playing short envelopes (sustain = 0) when the sound stops, the notes are still running.
for example, if you play a short pluck sound with no sustain volume but with long notes, it's still rendering all the voices even though no sound is generated.

this patch essentially say, if our sustain is 0.0 then when we've reached the end of the decay, we can skip to finished as release will have no affect.

double checked this with sustain of zero but slower decay and longer release and this causes the decay to be skipped early and move straight onto release on note off.