automation de-dupe
scans the automation points to find any where three in a row have the same value. if then removes the middle point and adds the time onto the first point.
this is done after the points have been delta and byte coded as the point values are more likely to be the same