convert all c++ sources to unix-newlines in repo
Git isn't really good at diffing CRLF files, and it's strongly
recommended to use LF files in the repository. Which EOL-style ends up
in the working-directory depends on the core.autocrlf and core.eol
settings in your git-config.

So let's encode C++ sources as LF instead of CRLF, as this will work
better on non-windows systems.

HACK: This commit was done by Erik Faye-Lund, but with ferris as the
author, as that reflects actual authorship much better. All I did was
mechanically convert sources.