WaveSabreVstLib: disable warning about deprecated functions
We don't control these source-files, so we don't really care if these
function names are deprecated or not.

There's no real chance of these functions being actually removed,
because literally everyone already uses them. But even if they were, we
would notice, and we could inject some macros instead to use the non
POSIX names if needed to.

But for now, let's keep this simple.