WaveSabreCore: suppress x64 cast-warning
OpenFile is a bit strange. It returns a HFILE instead of a
HANDLE-object, and on x64 these are of different sizes. Microsoft
recommends against using OpenFile, and says new applications should use
CreateFile instead.

But we're in 64k-land, and here we hack for size. OpenFile takes much
less arguments, so it generates shorter code to call it. So let's keep
it as-is, and just cast via an UINT_PTR to show the compiler that this
is really what we want.