WaveSabreVstLib: explicitly build static lib
WaveSabreVstLib has to be built as a static lib, because it contains the
VST code that references createEffect, which is only defined in the
libraries that reference it. So let's make that explicit.

This allows us to build with -DBUILD_SHARED_LIBS=ON as well. Not that
it's a particularly useful switch, but it's nice to be robust.