cmake: default to x86 builds on newer VS versions
WaveSabre is mainly an x86 project, and up until CMake 3.14 it always
picked the Win32 target for MSVC projects. But in version 3.14 this
changed when the VS 2019 generator was added, which picks x64 by default
on x64 machines.

So let's override that, so we compile x86 code by default. It's what
most users are going to want.