Simplify FastCos
The primary motivation is that the previous code relied on a 64-bit left shift, which with MSVC 14.2 and higher targeting x86, would generate code that relied on a function from MSVCRT.lib (not .dll) that we don't implement. By constraining the input range to [1, 2) using 1 + x - floor(x), we guarantee that the phase's exponent is always 0, so we can just take the significand bits without shifting them by the exponent. There's still a right-shift to isolate the significand, but this is by a constant factor, and is thus compiled without any issues.

Additionally, because we limit the range this way, we also don't need to take advantage of cosine's symmetry about the y axis, so we remove the abs call as well.

This new implementation has almost identical error compared to the previous, and is actually about 10% slower, but it works out-of-the-box on both x86 and x64 in the MinSizeRel configuration without any additional changes, and should still be quite portable.