Clean up duplicate fast cos tab size constants
We had these defined in two places, and the definitions were actually divergent. The table size ended up being determined by the size definitions in the cpp file, however, all indexing/usage of the table was determined by the Helpers members. Luckily, this actually behaves exactly as intended, but has the consequence that the table occupies twice as much memory as it should have (and the second half was never touched). This is super nasty, though, so let's clean it up.