ProjectManager: Remove reference to removed Track.DataSize field
This was removed in 271d6c924dfeab1a21c094e5086cd8c144ca8bdc in order to only have actual song data as part of the Song hierarchy, with no metadata. This is not only simpler logically, but is also important now that our converter tests automatically include all fields in the hierarchy, so anything other than relevant data should be removed.

Further, this size was pre-compression, so I'd argue it may not be very useful to begin with.